- Fixed ValveBlock WAILA tool tip showing 0% capacity for high fluid amounts.

+++ src/main/java/net/zarathul/simplefluidtanks/waila/ValveBlockDataProvider.java
@@ -53,7 +53,7 @@ public final class ValveBlockDataProvider implements IWailaDataProvider
 
 			if (config.getConfig(Registry.WAILA_TOTAL_CAPACITY_KEY))
 			{
-				int totalFillPercentage = MathHelper.clamp_int(valveEntity.getFluidAmount() * (100 / valveEntity.getCapacity()), 0, 100);
+				int totalFillPercentage = MathHelper.clamp_int((int)((long)valveEntity.getFluidAmount() * 100 / valveEntity.getCapacity()), 0, 100);
 				
 				if (config.getConfig(Registry.WAILA_CAPACITY_IN_MILLIBUCKETS_KEY))
 				{
